refactor(sstv2): fix logic bugs and apply clang-tidy cleanups
**Logic bug fixes:**
- block: eliminate deep copy in BlockBuilder::finish() by using std::move
- bloom: use full 64-bit hash (XOR-fold to 32-bit) to preserve entropy
- bloom: add overflow guard in bit_count calculation
- sstable: raise fixed-point iteration cap from 16 to 128 iterations

**Architecture:**
- Move block_flags (encode_block_flag, decode_block_flag) from
  compress/compress.h to format/block_flags.h

**Clang-tidy cleanups:**
- pattern/raw.h, types/value.h: replace std::enable_if_t with
  C++20 requires constraints (modernize-use-constraints)
- types/value.h: add default member initializer for type_
  (modernize-use-default-member-init)
- types/value.h: replace std::stable_sort with std::ranges::stable_sort
  (modernize-use-ranges)

**Formatting:**
- Apply clang-format across affected files

+++ cpp/pl/sstv2/bloom/bloom.h
@@ -53,7 +53,7 @@ public:
 
 private:
     int bits_per_key_;
-    std::vector<uint32_t> hashes_;
+    std::vector<uint64_t> hashes_;
 };
 
 class Reader {
@@ -61,8 +61,8 @@ public:
     [[nodiscard]] static absl::StatusOr<Reader> open(std::string_view section);
 
     [[nodiscard]] bool may_contain_all_key(const types::EncodedAllKey& all_key) const;
-    [[nodiscard]] absl::StatusOr<bool> may_contain(const types::InternalRow& row,
-                                                   const types::InternalSchema::ConstRef& schema) const;
+    [[nodiscard]] absl::StatusOr<bool> may_contain(
+        const types::InternalRow& row, const types::InternalSchema::ConstRef& schema) const;
 
     [[nodiscard]] const Header& header() const noexcept { return header_; }
